Inherit tests for OpenstackQueryEmailBase
